Move pipeline chain into pipeline::chain
