+++ gotham/src/router/builder/single.rs
@@ -1,11 +1,12 @@
 use std::panic::RefUnwindSafe;
 
 use extractor::{PathExtractor, QueryStringExtractor};
+use pipeline::chain::PipelineHandleChain;
 use router::builder::SingleRouteBuilder;
 use router::builder::replace::{ReplacePathExtractor, ReplaceQueryStringExtractor};
 use router::route::{Delegation, Extractors, RouteImpl};
 use router::route::matcher::RouteMatcher;
-use router::route::dispatch::{DispatcherImpl, PipelineHandleChain};
+use router::route::dispatch::DispatcherImpl;
 use handler::{Handler, NewHandler};
 
 /// Describes the API for defining a single route, after determining which request paths will be
