+++ gotham/src/router/route/dispatch.rs
@@ -2,12 +2,11 @@
 
 use std::sync::Arc;
 use std::panic::RefUnwindSafe;
-use borrow_bag::{BorrowBag, Handle, Lookup};
+use borrow_bag::BorrowBag;
 use futures::future;
 
 use handler::{Handler, HandlerFuture, IntoHandlerError, NewHandler};
-use middleware::chain::NewMiddlewareChain;
-use pipeline::Pipeline;
+use pipeline::chain::PipelineHandleChain;
 use state::{request_id, State};
 
 /// Represents the set of all `Pipeline` instances that are available for use with `Routes`.
@@ -91,62 +90,6 @@ where
                 Box::new(future::err((state, e.into_handler_error())))
             }
         }
-    }
-}
-
-/// A heterogeneous list of `Handle<P, _>` values, where `P` is a pipeline type. The pipelines are
-/// borrowed and invoked in order to serve a request.
-///
-/// Implemented using nested tuples, with `()` marking the end of the list. The list is in the
-/// reverse order of what is described via the routing API.
-///
-/// That is:
-///
-/// `(p3, (p2, (p1, ())))`
-///
-/// will be invoked as:
-///
-/// `(state, request)` &rarr; `p1` &rarr; `p2` &rarr; `p3` &rarr; `handler`
-pub trait PipelineHandleChain<P>: RefUnwindSafe {
-    /// Invokes this part of the `PipelineHandleChain`, with requests being passed through to `f`
-    /// once all `Middleware` in the `Pipeline` have passed the request through.
-    fn call<F>(&self, pipelines: &PipelineSet<P>, state: State, f: F) -> Box<HandlerFuture>
-    where
-        F: FnOnce(State) -> Box<HandlerFuture> + 'static;
-}
-
-/// Part of a `PipelineHandleChain` which references a `Pipeline` and continues with a tail element.
-impl<'a, P, T, N, U> PipelineHandleChain<P> for (Handle<Pipeline<T>, N>, U)
-where
-    T: NewMiddlewareChain,
-    T::Instance: 'static,
-    U: PipelineHandleChain<P>,
-    P: Lookup<Pipeline<T>, N>,
-    N: RefUnwindSafe,
-{
-    fn call<F>(&self, pipelines: &PipelineSet<P>, state: State, f: F) -> Box<HandlerFuture>
-    where
-        F: FnOnce(State) -> Box<HandlerFuture> + 'static,
-    {
-        let (handle, ref chain) = *self;
-        match pipelines.borrow(handle).construct() {
-            Ok(p) => chain.call(pipelines, state, move |state| p.call(state, f)),
-            Err(e) => {
-                trace!("[{}] error borrowing pipeline", request_id(&state));
-                Box::new(future::err((state, e.into_handler_error())))
-            }
-        }
-    }
-}
-
-/// The marker for the end of a `PipelineHandleChain`.
-impl<P> PipelineHandleChain<P> for () {
-    fn call<F>(&self, _: &PipelineSet<P>, state: State, f: F) -> Box<HandlerFuture>
-    where
-        F: FnOnce(State) -> Box<HandlerFuture> + 'static,
-    {
-        trace!("[{}] start pipeline", request_id(&state));
-        f(state)
     }
 }
